Tests: one class per file
Move the `ConsoleColorWithForceSupport` fixture used in the tests to a separate file.
