+++ tests/ConsoleColorTest.php
@@ -2,34 +2,8 @@
 namespace PHP_Parallel_Lint\PhpConsoleColor\Test;
 
 use PHP_Parallel_Lint\PhpConsoleColor\ConsoleColor;
+use PHP_Parallel_Lint\PhpConsoleColor\Test\Fixtures\ConsoleColorWithForceSupport;
 use PHPUnit\Framework\TestCase;
-
-class ConsoleColorWithForceSupport extends ConsoleColor
-{
-    private $isSupportedForce = true;
-
-    private $are256ColorsSupportedForce = true;
-
-    public function setIsSupported($isSupported)
-    {
-        $this->isSupportedForce = $isSupported;
-    }
-
-    public function isSupported()
-    {
-        return $this->isSupportedForce;
-    }
-
-    public function setAre256ColorsSupported($are256ColorsSupported)
-    {
-        $this->are256ColorsSupportedForce = $are256ColorsSupported;
-    }
-
-    public function are256ColorsSupported()
-    {
-        return $this->are256ColorsSupportedForce;
-    }
-}
 
 class ConsoleColorTest extends TestCase
 {
